temporal/scripts/unstructured/image-classification/tensorflow/convolutional-neural-network/requirements.txt
tensorflow
python-dotenv